requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
attrs==24.3.0
CacheControl==0.14.3
cachetools==6.2.1
certifi==2024.12.14
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==46.0.3
cycler==0.12.1
deprecation==2.1.0
fastapi==0.115.6
filelock==3.16.1
firebase_admin==7.1.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.5.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
greenlet==3.2.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
lap==0.5.12
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.1.0
networkx==3.4.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
pillow==11.1.0
postgrest==2.22.1
propcache==0.2.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==6.1.1
psycopg2-binary==2.9.11
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
PyYAML==6.0.2
realtime==2.22.1
requests==2.32.3
rsa==4.9.1
scipy==1.15.0
seaborn==0.13.2
setuptools==75.8.0
shapely==2.0.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.41.3
storage3==2.22.1
StrEnum==0.4.15
supabase==2.22.1
supabase-auth==2.22.1
supabase-functions==2.22.1
sympy==1.13.1
thop==0.1.1.post2209072238
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
twilio==9.4.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
tzlocal==5.3.1
ultralytics==8.3.59
ultralytics-thop==2.0.13
urllib3==2.5.0
uvicorn==0.34.0
websockets==15.0.1
yarl==1.22.0
